requirements.txt
streamlit
tensorflow==2.12.0
numpy
pillow
h5py
matplotlib